fix for NanoVDB on GPU

+++ barney/api/barney.cu
@@ -562,11 +562,11 @@ namespace barney_api {
     Context *context = checkGet(_context);
     std::shared_ptr<Data> data
       = context->createData(slot,dataType);
-    data->set(items,(int)numItems);
+    data->set(items,numItems);
     return (BNData)context->initReference(data);
   }
 
-      BARNEY_API
+  BARNEY_API
   void bnDataSet(BNData _data,
                  size_t numItems,
                  const void *items)
